+++ middleman-core/lib/middleman-more/core_extensions/default_helpers.rb
@@ -1,16 +1,20 @@
 if !defined?(::Padrino::Helpers)
-  require 'vendored-middleman-deps/padrino-core-0.11.4/lib/padrino-core/support_lite'
-  require 'vendored-middleman-deps/padrino-helpers-0.11.4/lib/padrino-helpers'
+  require 'vendored-middleman-deps/padrino-core-0.12.0/lib/padrino-core/support_lite'
+  require 'vendored-middleman-deps/padrino-helpers-0.12.0/lib/padrino-helpers'
+
+  # Don't fail on invalid locale, that's not what our current
+  # users expect.
+  ::I18n.enforce_available_locales = false
 end
 
 class Padrino::Helpers::OutputHelpers::ErbHandler
   # Force Erb capture not to use safebuffer
   def capture_from_template(*args, &block)
     self.output_buffer, _buf_was = '', self.output_buffer
-    captured_block = block.call(*args)
-    ret = eval('@_out_buf', block.binding)
+    raw = block.call(*args)
+    captured = template.instance_variable_get(:@_out_buf)
     self.output_buffer = _buf_was
-    [ ret, captured_block ]
+    engine_matches?(block) ? captured : raw
   end
 end
 
@@ -48,6 +52,7 @@ class Middleman::CoreExtensions::DefaultHelpers < ::Middleman::Extension
       attributes = tag_attributes(options)
       output = ActiveSupport::SafeBuffer.new
       output.safe_concat "<#{name}#{attributes}>"
+
       if content.respond_to?(:each) && !content.is_a?(String)
         content.each { |c| output.safe_concat c; output.safe_concat ::Padrino::Helpers::TagHelpers::NEWLINE }
       else
@@ -59,19 +64,18 @@ class Middleman::CoreExtensions::DefaultHelpers < ::Middleman::Extension
     end
 
     def capture_html(*args, &block)
-      handler = auto_find_proper_handler(&block)
-      captured_block, captured_html = nil, ''
-      if handler && handler.is_type? && handler.block_is_type?(block)
-        captured_html, captured_block = handler.capture_from_template(*args, &block)
+      if handler = auto_find_proper_handler(&block)
+        handler.capture_from_template(*args, &block)
+      else
+        block.call(*args)
       end
-      # invoking the block directly if there was no template
-      captured_html = block_given? && ( captured_block || block.call(*args) )  if captured_html.blank?
-      captured_html
     end
 
     def auto_find_proper_handler(&block)
       engine = block_given? ? File.extname(block.source_location[0])[1..-1].to_sym : current_engine
-      ::Padrino::Helpers::OutputHelpers.handlers.map { |h| h.new(self) }.find { |h| h.engines.include?(engine) && h.is_type? }
+      return if engine == :slim && engine == current_engine
+      handler_class = ::Padrino::Helpers::OutputHelpers.handlers[engine]
+      handler_class && handler_class.new(self)
     end
 
     # Disable Padrino cache buster