+++ middleman-core/lib/vendored-middleman-deps/padrino-core-0.12.0/test/fixtures/apps/helpers/system_helpers.rb
@@ -1,0 +1,8 @@
+require 'resolv'
+
+SystemDemo.helpers do
+  def resolv
+    Resolv.name
+  end
+end
+