+++ middleman-core/lib/vendored-middleman-deps/padrino-core-0.12.0/test/fixtures/apps/models/parent.rb
@@ -1,0 +1,5 @@
+class Parent
+  def family
+    'Danes'
+  end
+end