+++ middleman-core/lib/vendored-middleman-deps/padrino-core-0.12.0/test/fixtures/apps/system.rb
@@ -1,0 +1,13 @@
+PADRINO_ROOT = File.dirname(__FILE__) unless defined? PADRINO_ROOT
+
+class SystemDemo < Padrino::Application
+  set :reload, true
+end
+
+SystemDemo.controllers do
+  get '/' do
+    resolv
+  end
+end
+
+Padrino.load!