+++ middleman-core/lib/vendored-middleman-deps/padrino-helpers-0.12.0/lib/padrino-helpers/output_helpers/slim_handler.rb
@@ -1,0 +1,18 @@
+module Padrino
+  module Helpers
+    module OutputHelpers
+      ##
+      # Handler for Slim templates.
+      #
+      class SlimHandler < AbstractHandler
+        ##
+        # Returns true if the block is for Slim.
+        #
+        def engine_matches?(block)
+          block.binding.eval('defined? __in_slim_template')
+        end
+      end
+      OutputHelpers.register(:slim, SlimHandler)
+    end
+  end
+end